Moved error to dax package
